% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabtCI.R
\name{pppaccept}
\alias{pppaccept}
\title{Prior predictive acceptance probability}
\usage{
pppaccept(w, theta, mu, t2, s20, nu0, n, alpha)
}
\arguments{
\item{w}{the weight in [0,1] determining acceptance region endpoints}

\item{theta}{value of theta being tested}

\item{mu}{prior mean of theta}

\item{t2}{prior variance of theta}

\item{s20}{prior parameter for population variance}

\item{nu0}{prior parameter for population variance}

\item{n}{sample size}

\item{alpha}{level of test}
}
\description{
Compute prior probability of falling into an acceptance region
}
\details{
Internal function for fabtCI. This function computes the prior 
probability that the value theta will fall in the acceptance 
region of a biased level-alpha test, under a specified  prior 
predictive distribution.
}
\author{
Peter Hoff
}
\keyword{internal}

