% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radon}
\alias{radon}
\title{Minnesota Radon Dataset}
\format{A numeric matrix}
\source{
\url{http://www.stat.columbia.edu/~gelman/arm/software/}
}
\usage{
data(radon)
}
\description{
Radon levels in 919 homes from 85 Minnesota counties
}
\keyword{datasets}

