% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabregCI.R
\name{umauregCI}
\alias{umauregCI}
\title{UMAU regression coefficient intervals}
\usage{
umauregCI(y, X, alpha = 0.05)
}
\arguments{
\item{y}{a numeric vector of data}

\item{X}{a design matrix}

\item{alpha}{the type I error rate, so 1-alpha is the coverage rate}
}
\value{
A matrix where each row corresponds to the interval and OLS 
estimate of a coefficient.
}
\description{
Compute the usual t-intervals for the coefficients of a 
regression model
}
\details{
This function computes the 'usual' uniformly most 
accurate unbiased confidence interval for each coefficient in 
a linear regression model.
}
\author{
Peter Hoff
}

