% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabtCI.R
\name{wfabt}
\alias{wfabt}
\title{Bayes-optimal w-function}
\usage{
wfabt(theta, mu, t2, s20, nu0, n, alpha)
}
\arguments{
\item{theta}{value of theta being tested}

\item{mu}{prior mean of theta}

\item{t2}{prior variance of theta}

\item{s20}{prior parameter for population variance}

\item{nu0}{prior parameter for population variance}

\item{n}{sample size}

\item{alpha}{level of test}
}
\description{
Compute Bayes optimal w-function
}
\details{
Internal function for fabtCI. This function computes the 
value of w that minimizes the acceptance probability of a 
biased level-alpha test under a specified  prior
predictive distribution.
}
\author{
Peter Hoff
}
\keyword{internal}

