% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabzCI.R
\name{wfabz}
\alias{wfabz}
\title{Bayes-optimal w-function}
\usage{
wfabz(theta, mu, t2, s2, alpha = 0.05)
}
\arguments{
\item{theta}{value of theta being tested}

\item{mu}{prior mean of theta}

\item{t2}{prior variance of theta}

\item{s2}{population variance}

\item{alpha}{level of test}
}
\description{
Compute Bayes optimal w-function
}
\details{
This function computes the 
value of w that minimizes the acceptance probability of a 
biased level-alpha test for a normal population with 
known variance, under a specified  prior
predictive distribution.
}
\author{
Peter Hoff
}
\keyword{internal}

